dismiss on admin

+++ lib/src/features/home/presentation/dismiss_me.dart
@@ -3,6 +3,7 @@ import 'package:firebase_ui_firestore/firebase_ui_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import '../../../utils/constants/constants.dart';
 import '../../../utils/providers.dart';
 import '../data/medicine.dart';
 import '../hive_functions/ambulance_provider.dart';
@@ -26,21 +27,58 @@ class DismissMe extends ConsumerWidget {
         final medicine = doc.data();
         return Dismissible(
           key: Key(doc.id),
+          confirmDismiss: (_) async {
+            return await showDialog(
+              context: context,
+              builder: (BuildContext context) {
+                return AlertDialog(
+                  title: Text(medicine.name),
+                  content: const SingleChildScrollView(
+                    child: ListBody(
+                      children: <Widget>[
+                        Text(AppStrings.deleteOne),
+                      ],
+                    ),
+                  ),
+                  actions: <Widget>[
+                    TextButton(
+                      child: const Text(AppStrings.cancel),
+                      onPressed: () {
+                        Navigator.of(context).pop();
+                      },
+                    ),
+                    TextButton(
+                      child: const Text(AppStrings.yes),
+                      onPressed: () async {
+                        final user = ref.read(firebaseAuthProvider).currentUser;
+                        ref
+                            .read(firestoreRepositoryProvider)
+                            .deleteMedicine(user!.uid, doc.id);
+                        Navigator.of(context).pop();
+                      },
+                    ),
+                  ],
+                );
+              },
+            );
+          },
           background: const ColoredBox(color: Colors.red),
           direction: DismissDirection.endToStart,
           onDismissed: (direction) {
-            final user = ref.read(firebaseAuthProvider).currentUser;
-            ref
-                .read(firestoreRepositoryProvider)
-                .deleteMedicine(user!.uid, doc.id);
+            ScaffoldMessenger.of(context).showSnackBar(
+              SnackBar(content: Text("${medicine.name} - usunięte")),
+            );
           },
           child: ListTile(
             title: medicine.isDrug
                 ? Text(
                     medicine.name,
-                    style: const TextStyle(color: Colors.red),
+                    style: AppTextStyle.listTileRedFont,
                   )
-                : Text(medicine.name),
+                : Text(
+                    medicine.name,
+                    style: AppTextStyle.listTileFont,
+                  ),
             subtitle: medicine.createdAt != null
                 ? Text(medicine.createdAt.toString(),
                     style: Theme.of(context).textTheme.bodySmall)
